# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Event-driven scalable notification system API.",
        "title": "Insider Notification API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/notifications": {
            "get": {
                "description": "Filtreleme ve Keyset Cursor destekli yüksek performanslı bildirim listesi getirir.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Bildirimleri Listele",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Duruma göre filtrele (pending, processing, delivered, failed_retrying, failed_permanently, canceled)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Kanala göre filtrele (sms, email, push)",
                        "name": "channel",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Başlangıç tarihi filtresi (ISO8601 veya YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Bitiş tarihi filtresi (ISO8601 veya YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Sayfa boyutu (varsayılan 10, max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Bir sonraki sayfa için imleç (next_cursor)",
                        "name": "cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Transactional Outbox pattern kullanarak asenkron tekil bildirim işler.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Tekil bildirim oluşturur",
                "parameters": [
                    {
                        "description": "Bildirim Bilgileri",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.NotificationRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Bildirim alindi, isleniyor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Gecersiz JSON veya bilinmeyen alan",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Idempotency key cakismasi",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "429": {
                        "description": "Rate limit asildi",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Veritabanina yazilamadi",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/notifications/batch": {
            "post": {
                "description": "Transactional Outbox pattern kullanarak asenkron toplu bildirim (maksimum 1000 adet) işler.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Toplu bildirim oluşturur",
                "parameters": [
                    {
                        "description": "Bildirim Listesi",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.NotificationRequest"
                            }
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Toplu bildirimler alindi, isleniyor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Gecersiz JSON array formatı, bos liste veya limit asimi",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Idempotency key cakismasi",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "429": {
                        "description": "Rate limit asildi",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Toplu veritabani yazma islemi basarisiz",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/notifications/batch/{id}": {
            "get": {
                "description": "Batch ID'ye ait tüm bildirimlerin durumlarını listeler.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Toplu Bildirim Durumunu Getir",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Batch ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.NotificationResponse"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.NotificationRequest": {
            "type": "object",
            "properties": {
                "channel": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "idempotency_key": {
                    "type": "string"
                },
                "priority": {
                    "type": "string"
                },
                "recipient": {
                    "type": "string"
                }
            }
        },
        "models.NotificationResponse": {
            "type": "object",
            "properties": {
                "batch_id": {
                    "type": "string"
                },
                "channel": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "external_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "idempotency_key": {
                    "type": "string"
                },
                "next_retry_at": {
                    "type": "string"
                },
                "priority": {
                    "type": "string"
                },
                "recipient": {
                    "type": "string"
                },
                "retry_count": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}